Use Colorwheel for color selection
Replace the custom RGB(W) sliders with an iro.js color wheel.
For RGBW LEDs, the alpha channel is used to control the white LED.

Foreground and background colors can be selected in the same color
picker next to each other.
A checkbox is added to quickly enable/disable the background color.